requirements.txt
playwright>=1.28.0
playwright-stealth>=0.0.1
python-dotenv>=0.21.0
